+++ packages/site/src/pages/playground.tsx
@@ -62,11 +62,10 @@ const Playground = () => {
   const [nocturneFrontendSDK, setFrontendSDK] = useState<NocturneFrontendSDK>();
 
   useEffect(() => {
-    loadNocturneFrontendSDK(
-      DEPOSIT_MANAGER_CONTRACT_ADDRESS,
-      BUNDLER_ENDPOINT,
-      SCREENER_ENDPOINT
-    ).then((sdk) => {
+    loadNocturneFrontendSDK(DEPOSIT_MANAGER_CONTRACT_ADDRESS, {
+      screenerEndpoint: SCREENER_ENDPOINT,
+      bundlerEndpoint: BUNDLER_ENDPOINT,
+    }).then((sdk) => {
       console.log("instantiated frontend sdk");
       setFrontendSDK(sdk);
     });