+++ packages/site/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- fix mixup between screener and bundler endpoints
 - add DepositForm to test site for fe-sdk testing
 - Pass in `chainid` and `deadline` to operation request in test site
 - Remove hardcoded vault address after wallet <> handler separation